requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
numpy==1.26.4
ordered-set==4.1.0
pdf2image==1.17.0
pillow==10.3.0
PyLaTeX==1.4.2
python-dotenv==1.0.1
yarl==1.9.4
